refactor: Request/Response 구조체 파일 분리
