+++ iOS_Wegg/iOS_Wegg/Core/Services/API/Targets/Map/HotPlacesAPI.swift
@@ -8,37 +8,6 @@
 import Foundation
 import Moya
 
-struct HotPlaceRequest: Encodable {
-    let minX: Double
-    let maxX: Double
-    let minY: Double
-    let maxY: Double
-    let sortBy: String
-}
-
-struct HotPlacesResponse: Codable {
-    let isSuccess: Bool
-    let code: String
-    let message: String
-    let result: HotPlacesResult
-    
-    struct HotPlacesResult: Codable {
-        let hotPlaceList: [HotPlace]
-    }
-    
-    struct HotPlace: Codable {
-        let addressId: Int
-        let latitude: Double
-        let longitude: Double
-        let placeName: String
-        let placeLabel: String
-        let authCount: Int
-        let saveCount: Int
-        let distance: Double
-        let postList: [String] // 현재 postList는 비어 있으므로 String 배열로 처리
-    }
-}
-
 enum HotPlacesAPI {
     case getHotPlaces(request: HotPlaceRequest)
 }
