+++ iOS_Wegg/iOS_Wegg/Models/NetworkModel/Schedule/ScheduleSearchResponse.swift
@@ -1,17 +1,11 @@
 //
-//  ScheduleSearch.swift
+//  ScheduleSearchResponse.swift
 //  iOS_Wegg
 //
 //  Created by jaewon Lee on 2/10/25.
 //
 
 import Foundation
-
-struct ScheduleSearchRequest: Encodable {
-    let keyword: String
-    let latitude: String
-    let longitude: String
-}
 
 struct ScheduleSearchResponse: Decodable {
     let isSuccess: Bool
